Moved InvocationAnalyzer where it belongss
